requirements.txt
chardet
fpdf
nltk
numpy
beautifulsoup4
torch
reportlab
Pypdf2
cairosvg
svgwrite
weasyprint
pdfkit
PyMuPDF
python-docx
mpdf
xlsxwriter
openpyxl
pymsgbox
transformers
matplotlib
markdown2pdf